+++ R/manyhealth-actors.R
@@ -1,0 +1,11 @@
+#' actors database documentation
+#'
+#' @format The actors database is a list that contains the
+#' following 1 datasets: CHATHAM.
+#' For more information and references to each of the datasets used,
+#' please use the `data_source()` and `data_contrast()` functions.
+#'\describe{
+#' \item{CHATHAM: }{A dataset with 203 observations and the following
+#' 5 variables: Actor, ID, Beg, Country, City.}
+#' }
+ "actors"